Ensure organization aware IdentityProviderModel is used in the infinispan IDPProvider
Closes #32108
